DEV: Convert share-topic modal to new component-based API
