+++ app/assets/javascripts/discourse/app/components/composer-messages.hbs
@@ -2,7 +2,13 @@
   <ComposerMessage
     @message={{message}}
     @closeMessage={{action "closeMessage"}}
-    @shareModal={{action "shareModal"}}
+    @shareModal={{fn (mut this.showShareModal true)}}
     @switchPM={{action "switchPM"}}
   />
+  {{#if this.showShareModal}}
+    <Modal::ShareTopic
+      @closeModal={{fn (mut this.showShareModal) false}}
+      @model={{this.shareModalData}}
+    />
+  {{/if}}
 {{/each}}
