+++ app/assets/javascripts/discourse/app/instance-initializers/topic-footer-buttons.js
@@ -4,7 +4,7 @@ import {
   WITH_REMINDER_ICON,
 } from "discourse/models/bookmark";
 import { registerTopicFooterButton } from "discourse/lib/register-topic-footer-button";
-import showModal from "discourse/lib/show-modal";
+import ShareTopicModal from "discourse/components/modal/share-topic";
 
 const SHARE_PRIORITY = 1000;
 const BOOKMARK_PRIORITY = 900;
@@ -13,7 +13,7 @@ const FLAG_PRIORITY = 700;
 const DEFER_PRIORITY = 500;
 
 export default {
-  initialize() {
+  initialize(owner) {
     registerTopicFooterButton({
       id: "share-and-invite",
       icon: "d-topic-share",
@@ -25,15 +25,15 @@ export default {
       },
       title: "topic.share.help",
       action() {
-        const controller = showModal("share-topic", {
-          model: this.topic.category,
-        });
-        controller.setProperties({
-          allowInvites:
-            this.currentUser.can_invite_to_forum &&
-            this.canInviteTo &&
-            !this.inviteDisabled,
-          topic: this.topic,
+        owner.lookup("service:modal").show(ShareTopicModal, {
+          model: {
+            category: this.topic.category,
+            topic: this.topic,
+            allowInvites:
+              this.currentUser.can_invite_to_forum &&
+              this.canInviteTo &&
+              !this.inviteDisabled,
+          },
         });
       },
       dropdown() {
